Add first support for OCF scenes.
This commit adds a first support for OCF scenes with certain limitations:
- No support for remote setup of scenes. Can only be created via local API.
- Scenes are currently bound to device 0.
- When a scene is triggered, then the device values are not updated.
  Instead the app receives a callback to handle it itself.
- Not compliant with IoTivity as that is using wrong interfaces (oic.if.b
  instead of oic.if.a as specified in OCF).

Reviewed-on: https://gerrit.iotivity.org/gerrit/21883
Tested-by: Reinhard Baier <baier@lynxtechnology.com>

+++ include/oc_rep.h
@@ -233,8 +233,8 @@ typedef struct oc_rep_s
 
 void oc_rep_set_pool(struct oc_memb *rep_objects_pool);
 
-uint16_t oc_parse_rep(const uint8_t *payload, uint16_t payload_size,
-                      oc_rep_t **value_list);
+int oc_parse_rep(const uint8_t *payload, uint16_t payload_size,
+                 oc_rep_t **value_list);
 
 void oc_free_rep(oc_rep_t *rep);
